+++ rsconnect/main.py
@@ -1138,7 +1138,7 @@ def _warn_on_ignored_requirements(directory: str, requirements_file_name: str):
     default="requirements.txt",
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -1319,7 +1319,7 @@ def deploy_notebook(
     default="requirements.txt",
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -1743,7 +1743,7 @@ def deploy_pyproject(
     default="requirements.txt",
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -2173,7 +2173,7 @@ def generate_deploy_python(app_mode: AppMode, alias: str, min_version: str, desc
         type=click.Path(dir_okay=False),
         help=(
             "Path to requirements file listing the project dependencies. "
-            "Any file compatible with requirements.txt format or uv.lock is accepted, "
+            "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
             "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
             "Must be inside the project directory."
         ),
@@ -2540,7 +2540,7 @@ def write_manifest():
     type=click.Path(dir_okay=False),
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -2659,7 +2659,7 @@ def write_manifest_notebook(
     type=click.Path(exists=True, dir_okay=False),
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -2814,7 +2814,7 @@ def write_manifest_voila(
     type=click.Path(dir_okay=False),
     help=(
         "Path to requirements file listing the project dependencies. "
-        "Any file compatible with requirements.txt format or uv.lock is accepted, "
+        "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
         "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
         "Must be inside the project directory."
     ),
@@ -3031,7 +3031,7 @@ def generate_write_manifest_python(app_mode: AppMode, alias: str, desc: Optional
         type=click.Path(dir_okay=False),
         help=(
             "Path to requirements file listing the project dependencies. "
-            "Any file compatible with requirements.txt format or uv.lock is accepted, "
+            "Any file compatible with requirements.txt format, uv.lock, or pyproject.toml is accepted, "
             "a requirements.txt.lock retrieved with 'rsconnect content get-lockfile' is also supported. "
             "Must be inside the project directory."
         ),